Made resolvers modular and added to api get animal by id
