+++ back-end/server/src/schema/Animal.graphql
@@ -1,4 +1,5 @@
 type Animal {
+    id: String!
     name: String!
     type: String!
     dateOfBirth: String
