+++ back-end/server/src/schema/Query.graphql
@@ -1,3 +1,4 @@
 type Query {
     animals: [Animal]
+    animal(id: String!): Animal
 }
