+++ front-end/src/components/AnimalsList.vue
@@ -21,6 +21,7 @@ export default {
     setup() {
         const animals = ref([
             {
+                id: '1',
                 name: 'Kidney',
                 type: 'Cat',
                 dateOfBirth: new Date('2005-05-05').toLocaleDateString(),
@@ -28,6 +29,7 @@ export default {
                     'https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcSlNyI5Bbsl1vq1BQjH9XA-Z4j0Kkk0cEpAnA&usqp=CAU',
             },
             {
+                id: '2',
                 name: 'Kfitz',
                 type: 'Cat',
                 dateOfBirth: new Date('2007-05-05').toLocaleDateString(),
@@ -35,6 +37,7 @@ export default {
                     'https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcSlNyI5Bbsl1vq1BQjH9XA-Z4j0Kkk0cEpAnA&usqp=CAU',
             },
             {
+                id: '3',
                 name: 'Peter',
                 type: 'Cat',
                 dateOfBirth: new Date('2010-05-05').toLocaleDateString(),
@@ -42,6 +45,7 @@ export default {
                     'https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcSlNyI5Bbsl1vq1BQjH9XA-Z4j0Kkk0cEpAnA&usqp=CAU',
             },
             {
+                id: '4',
                 name: 'Pizi',
                 type: 'Cat',
                 dateOfBirth: new Date('2012-05-05').toLocaleDateString(),
@@ -49,6 +53,7 @@ export default {
                     'https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcSlNyI5Bbsl1vq1BQjH9XA-Z4j0Kkk0cEpAnA&usqp=CAU',
             },
             {
+                id: '5',
                 name: 'Alona',
                 type: 'Cat',
                 dateOfBirth: new Date('2014-05-05').toLocaleDateString(),
@@ -56,6 +61,7 @@ export default {
                     'https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcSlNyI5Bbsl1vq1BQjH9XA-Z4j0Kkk0cEpAnA&usqp=CAU',
             },
             {
+                id: '6',
                 name: 'Andey',
                 type: 'Cat',
                 dateOfBirth: new Date('2016-05-05').toLocaleDateString(),
@@ -63,6 +69,7 @@ export default {
                     'https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcSlNyI5Bbsl1vq1BQjH9XA-Z4j0Kkk0cEpAnA&usqp=CAU',
             },
             {
+                id: '7',
                 name: 'Libi',
                 type: 'Cat',
                 dateOfBirth: new Date('2018-05-05').toLocaleDateString(),
@@ -70,6 +77,7 @@ export default {
                     'https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcSlNyI5Bbsl1vq1BQjH9XA-Z4j0Kkk0cEpAnA&usqp=CAU',
             },
             {
+                id: '8',
                 name: 'Pumi',
                 type: 'Cat',
                 dateOfBirth: new Date('2019-05-05').toLocaleDateString(),
@@ -77,6 +85,7 @@ export default {
                     'https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcSlNyI5Bbsl1vq1BQjH9XA-Z4j0Kkk0cEpAnA&usqp=CAU',
             },
             {
+                id: '9',
                 name: 'Haim',
                 type: 'Cat',
                 dateOfBirth: new Date('2020-05-05').toLocaleDateString(),
